pyproject.toml
[project]
name = "discord-soundboard"
version = "0.1.0"
description = "Discord soundboard application with web dashboard"
requires-python = ">=3.8"
dependencies = [
    "discord.py==2.3.2",
    "aiohttp==3.9.1",
    "python-dotenv==1.0.0",
    "pynacl==1.5.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bot"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "black>=22.0.0",
    "flake8>=4.0.0"
]